Move get_file_object tasks to shared_tasks
