+++ datascience/src/pipeline/flows/regulations_open_data.py
@@ -1,5 +1,3 @@
-from io import BytesIO
-
 import geopandas as gpd
 import pandas as pd
 from prefect import Flow, Parameter, case, task
@@ -11,12 +9,14 @@ from config import (
     REGULATIONS_DATASET_ID,
     REGULATIONS_GEOPACKAGE_RESOURCE_ID,
     REGULATIONS_GEOPACKAGE_RESOURCE_TITLE,
-    ROOT_DIRECTORY,
 )
 from src.pipeline.generic_tasks import extract
 from src.pipeline.shared_tasks.control_flow import check_flow_not_running
-from src.pipeline.shared_tasks.datagouv import update_resource
-from src.pipeline.utils import remove_file
+from src.pipeline.shared_tasks.datagouv import (
+    get_csv_file_object,
+    get_geopackage_file_object,
+    update_resource,
+)
 
 
 @task(checkpoint=False)
@@ -55,104 +55,6 @@ def get_regulations_for_geopackage(regulations: gpd.GeoDataFrame) -> gpd.GeoData
     ]
 
     return regulations[columns].copy(deep=True)
-
-
-@task(checkpoint=False)
-def get_csv_file_object(df: pd.DataFrame) -> BytesIO:
-    """
-    Returns a `BytesIO` csv file object from the input `DataFrame`.
-    Useful to upload a `DataFrame` to data.gouv.fr
-
-    The index is not included in the output csv file.
-
-    Args:
-        df (pd.DataFrame): DataFrame to convert
-
-    Returns:
-        BytesIO: file object
-
-    Examples:
-        import pandas as pd
-        >>> df = pd.DataFrame({"a": [10, 20, 30], "b": [40, 50, 60]})
-        >>> df
-                a   b
-            0  10  40
-            1  20  50
-            2  30  60
-        >>> buf = df_to_csv_file_object.run(df)
-        >>> pd.read_csv(buf)
-                a   b
-            0  10  40
-            1  20  50
-            2  30  60
-    """
-    buf = BytesIO()
-    df.to_csv(buf, mode="wb", index=False)
-    buf.seek(0)
-    return buf
-
-
-@task(checkpoint=False)
-def get_geopackage_file_object(gdf: gpd.GeoDataFrame, layers: str = None) -> BytesIO:
-    """
-    Returns a `BytesIO` geopackage file object. from the input `GeoDataFrame`.
-
-    If `layers` is given, the geopackage will be organized in layers according to the
-    data labels of the `layers` column. If there are null values in the `layers` column,
-    the corresponding rows will not be included in the geopackage.
-
-    Args:
-        gdf (gpd.DataFrame): GeoDataFrame to convert
-        layers (str, optional): name of the column to use as layer labels in the
-          geopackage. Defaults to None.
-
-    Returns:
-        BytesIO: file object
-
-    Examples:
-        import geopandas as gpd
-        from shapely.geometry import Point
-        >>> gdf = gpd.GeoDataFrame({
-            "a": [10, 20, 30],
-            "geometry": [Point(1, 2), Point(3, 4), Point(5, 6)]
-        })
-        >>> gdf
-                a                 geometry
-            0  10  POINT (1.00000 2.00000)
-            1  20  POINT (3.00000 4.00000)
-            2  30  POINT (5.00000 6.00000)
-        >>> buf = get_geopackage_file_object.run(gdf)
-        >>> gpd.read_file(buf, driver="GPKG")
-                a                 geometry
-            0  10  POINT (1.00000 2.00000)
-            1  20  POINT (3.00000 4.00000)
-            2  30  POINT (5.00000 6.00000)
-    """
-
-    buf = BytesIO()
-
-    if layers:
-        temp_file_path = ROOT_DIRECTORY / "src/pipeline/data/tmp_geopackage.gpkg"
-        remove_file(temp_file_path, ignore_errors=True)
-
-        try:
-            for layer in gdf[layers].dropna().unique():
-
-                gdf[gdf[layers] == layer].to_file(
-                    temp_file_path, driver="GPKG", layer=layer
-                )
-
-            with open(temp_file_path, "rb") as f:
-                buf.write(f.read())
-
-        finally:
-            remove_file(temp_file_path, ignore_errors=True)
-
-    else:
-        gdf.to_file(buf, driver="GPKG")
-
-    buf.seek(0)
-    return buf
 
 
 with Flow("Regulations open data", executor=LocalDaskExecutor()) as flow:
